feat(main): implement lazy loading for tool widgets to improve startup performance
Migrate from eagerly loading all tool widgets at application startup to lazy loading them on-demand. This significantly improves startup time by deferring expensive imports until the user actually selects a specific tool.

Key changes:
- Removed eager creation of all 28 tool widgets in DevDriverWindow.__init__
- Added lazy_loader.py module with on-demand tool module importing
- Implemented caching mechanism for tool widgets and factory functions
- Updated main.py to use lazy loading with _get_or_create_tool_widget method
- Modified tools/__init__.py to expose only the lazy loader functions
- Added hidden imports to main.spec for PyInstaller compatibility

The Scratch Pad widget is still created eagerly as it's required for the dock system, but all other tools are now loaded only when selected by the user.

+++ devboost/tools/block_editor/formatters.py
@@ -104,7 +104,7 @@ def _format_value(v: CalcValue) -> str:
     if v.kind == "percent":
         return f"{_fmt_num(v.amount * 100)}%"
     if v.kind == "duration":
-        total = int(round(v.amount))
+        total = round(v.amount)
         h = total // 60
         m = total % 60
         if h > 0:
@@ -115,7 +115,7 @@ def _format_value(v: CalcValue) -> str:
 
 def _fmt_num(n: float) -> str:
     if abs(n - round(n)) < 1e-9:
-        return str(int(round(n)))
+        return str(round(n))
     return f"{n:.4f}".rstrip("0").rstrip(".")
